socket: resolve implicit conversion warnings flagged by codechecker
This patch addresses implicit conversion warnings reported by CodeChecker.
It modifies the code to use appropriate integer types or casts values to
the correct types, ensuring that conversions are explicit and safe.

This change should prevent potential issues related to implicit type
conversions and satisfies CodeChecker’s analysis.

Thanks to Moritz Buhl for doing codechecker testing.

+++ modules/include/uapi/linux/quic.h
@@ -129,7 +129,7 @@ enum quic_cong_algo {
 };
 
 struct quic_errinfo {
-	uint64_t stream_id;
+	int64_t  stream_id;
 	uint32_t errcode;
 };
 
@@ -176,8 +176,8 @@ enum {
 };
 
 struct quic_stream_update {
-	uint64_t id;
-	uint32_t state;
+	int64_t  id;
+	uint8_t  state;
 	uint32_t errcode;
 	uint64_t finalsz;
 };
